gradient-python/gradient-service-inference/requirements/requirements.txt
invoke==1.6.0
attrs==21.2.0
toolz==0.11.1
bidict==0.21.2
networkx==2.6.3
cloudpickle==1.6.0
protobuf==3.18.1
grpcio==1.33.2
pymongo==3.11.0
pandas==1.1.5
torch==1.9.1
pip==21.3
pytest==6.2.1 # dev
pytest-cov==2.10.1 # dev
hypothesis==6.0.0 # dev
pyyaml==5.4.1
gradient-service-domain==0.1.0+6
wheel==0.37.0
setuptools==58.2.0